fixed cyclic import and incorrect auto refactor
