+++ custom_components/polygonal_zones/services/delete_zone.py
@@ -4,11 +4,13 @@ import json
 from pathlib import Path
 
 from homeassistant.core import HomeAssistant, ServiceCall
-
-from custom_components.polygonal_zones.services.errors import ZoneDoesNotExists, ZoneFileNotEditable
+from .helpers import get_entities_from_device_id, get_zone_idx
+from .errors import (
+    ZoneDoesNotExists,
+    ZoneFileNotEditable,
+)
 from ..utils.general import load_data
 from ..utils.local_zones import save_zones
-from . import get_entities_from_device_id, get_zone_idx
 
 
 def delete_zone_action_builder(hass: HomeAssistant):
